requirements.txt
tensorflow>=2.13,<2.18
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
Pillow>=10.0
matplotlib>=3.7
streamlit>=1.36
